Clean up

+++ DarniProject/Assets/Scripts/ApplyEffects/AddDashChargeSlot.cs
@@ -1,8 +1,8 @@
 ﻿using UnityEngine;
 public class AddDashChargeSlot : ApplyEffect
 {
-    public override void Apply(PlayerXP player, int value)
+    public override void Apply(PlayerXP player, float value)
     {
-        player.movementStats.AddDashChargeSlots(value);
+        player.movementStats.AddDashChargeSlots((int)value);
     }
 }
